Make conditional helpers/references not volatile
We previously marked them as `volatile` with a FIXME because we need to
support proxies. This reduces the scope and only mark conditionals with
proxies as volatile (we can still do better than that by making proxy
objects have a special tag that combines its own `DirtyableTag` with the
tag from its `content`, but it's pretty low priority at the moment).

The refactor also added a new optimizations: by reusing the same code
across the syntax (`{{#if}}`) and the helper (`{{if}}`) versions, the
inline conditional helpers now also participate in the same const
optimization, i.e. `{{if true foo bar}}` will be optimized into `{{foo}}`,
so as `{{if @stuff foo bar}}` if `@stuff` is static.


TODO: fix tests that uses fake proxies, see comment in `references.js`